+++ inc/console/exception/earlyexitexception.class.php
@@ -30,42 +30,19 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Console\Database;
+namespace Glpi\Console\Exception;
 
 if (!defined('GLPI_ROOT')) {
    die("Sorry. You can't access this file directly");
 }
 
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Exception\ExceptionInterface;
 
-class ConfigureCommand extends AbstractConfigureCommand {
+/**
+ * This exception is used to easilly trigger an exit of current command from a sub method.
+ *
+ * @since x.x.x
+ */
+class EarlyExitException extends \Exception implements ExceptionInterface {
 
-   protected function configure() {
-
-      parent::configure();
-
-      $this->setName('glpi:database:configure');
-      $this->setAliases(['db:configure']);
-      $this->setDescription('Define database configuration');
-
-      $this->addOption(
-         'use-utf8mb4',
-         null,
-         InputOption::VALUE_NONE,
-         __('Use utf8mb4 character set.')
-      );
-   }
-
-   protected function execute(InputInterface $input, OutputInterface $output) {
-
-      $result = $this->configureDatabase($input, $output, $input->getOption('use-utf8mb4'));
-
-      if (self::ABORTED_BY_USER === $result) {
-         return 0; // Considered as success
-      }
-
-      return $result;
-   }
 }